Remote transcoding now mostly working

+++ transcode.sh
@@ -5,10 +5,10 @@
 #   $2-$n: Files inside $1 to transcode
 
 exec &> transcode.log
-echo "Beginning transcode"
 
 # Random ASCII name for the droplet
-DROPLET_NAME=$(uuidgen | shasum | awk '{print $1}')
+#DROPLET_NAME=$(uuidgen | shasum | awk '{print $1}')
+DROPLET_NAME='transcode-droplet'
 
 # Create new DigitalOcean compute droplet...
 #
@@ -18,11 +18,11 @@ DROPLET_NAME=$(uuidgen | shasum | awk '{print $1}')
 # Benchmark times:
 #   c-16 22m: 5m 20s (~5.5x) ~$0.5/hr
 #   c-32 22m: 5m 10s (~6.7x) ~$1/hr
-doctl compute droplet create "$DROPLET_NAME" \
-      --size c-8 \
-      --image ubuntu-16-04-x64 \
-      --region lon1 \
-      --ssh-keys 23354048
+#doctl compute droplet create "$DROPLET_NAME" \
+#      --size c-8 \
+#      --image ubuntu-16-04-x64 \
+#      --region lon1 \
+#      --ssh-keys 23354048
 
 # Its associated ID is needed for certain doctl commands
 DROPLET_ID=$(
@@ -30,8 +30,9 @@ DROPLET_ID=$(
          --output json | jq -r '.[0].id'
 )
 
+echo "Got $DROPLET_ID from DigitalOcean"
+
 # Get the IP address of the newly-created droplet
-echo "Attempting to obtain droplet IP address..."
 SERVER_ADDRESS='null'
 
 # Wait for the server to actually have an IP address
@@ -42,19 +43,17 @@ do
       doctl compute droplet get $DROPLET_ID --output json | \
       jq -r '.[0].networks.v4[0].ip_address'
    )
-   echo "Got $SERVER_ADDRESS from doctl"
 done
 
-echo "IP address is $SERVER_ADDRESS"
-echo "Waiting for sshd to init... "
+echo "Server address is $SERVER_ADDRESS"
 
 # Get the public key of the new server
 ssh-keyscan -H $SERVER_ADDRESS >> ~/.ssh/known_hosts
 
-# Stop strict host key checking for the droplet (‘:’ means do nothing)
-ssh -N -o "StrictHostKeyChecking no" root@$SERVER_ADDRESS
+# Stop strict host key checking for the droplet
+ssh -o 'StrictHostKeyChecking no' root@$SERVER_ADDRESS ':'
 
-echo "Copying binaries to remote..."
+echo "Copying binaries..."
 
 # Quickest method to get binaries across to the transcoding server
 scp /usr/bin/ffmpeg /usr/bin/ffprobe /usr/bin/bc root@$SERVER_ADDRESS:/usr/bin
@@ -62,35 +61,34 @@ scp /usr/bin/ffmpeg /usr/bin/ffprobe /usr/bin/bc root@$SERVER_ADDRESS:/usr/bin
 # Copy media and hls creation script
 scp ~/create-hls-stream.sh ~/ffmpeg-monitor-progress.sh root@$SERVER_ADDRESS:~
 
+echo "Finished copying binaries"
+
+# Files used in the transcoding process
 STATUS_FILE="$1/status.json"
+TRANSCODE_PROGRESS_PIPE=transcode-progress
 
 # Ensure the droplet is always deleted no matter the script result
 # (shit's too damn expensive lolz) and remove the lockfiles
 cleanup() {
-   echo "Trapped exit signal!"
-   
-   if [ -e $STATUS_FILE.lock ]
-   then
-      echo "Removing lock file..."
-      rm $STATUS_FILE.lock
-   fi
-   
-   echo "Removing ssh pipe..."
-   unlink transcode-progress
-   
-   echo "Removing droplet '$DROPLET_NAME'"
-   doctl compute droplet delete -f $DROPLET_NAME
+   [ -e $STATUS_FILE.lock ] && rm $STATUS_FILE.lock
+   [ -p $TRANSCODE_PROGRESS_PIPE ] && rm $TRANSCODE_PROGRESS_PIPE
+   #doctl compute droplet delete -f $DROPLET_NAME
 }
 
 trap cleanup EXIT HUP INT QUIT TERM
 
-#
 # Appends a given key–value pair to the given
 # object located inside $STATUS_FILE
+#
+# Arguments:
+#  $1 -> URL of object to update
+#  $2 -> Key of object
+#  $3 -> Value of object as a JSON literal
 #
 # Example:
 #   update_status_field <url> "duration" 1200
 update_status_field() {
+   flock $STATUS_FILE.lock \
    jq --arg url "$1" \
       --arg key "$2" \
       --argjson val "$3" \
@@ -102,25 +100,35 @@ update_status_field() {
    sponge $STATUS_FILE
 }
 
+# Updates status.json with progress from ffmpeg
+# Arguments:
+#  $1 -> downloaded URL
+#  $2 -> media duration in seconds
 update_progress() {
-   while true
+   while [ -p $TRANSCODE_PROGRESS_PIPE ]
    do
-      if read line < transcode-progress; then
+      if read line; then
          # Progress lines are written back in the form
          # <SPEED> <SECONDS>
-         SPEED=$(echo "$line" | awk '{print $1}')
-         SECONDS=$(echo "$line" | awk '{print $2}')
+         PROGRESS_SPEED=$(echo "$line" | awk '{print $1}')
+         PROGRESS_SECONDS=$(echo "$line" | awk '{print $2}')
          
-         echo "Read '$line' from ffmpeg-progress"
-         echo "$line" >> progress.txt
+         PERCENT_COMPLETE=$(
+            echo "scale=3; $PROGRESS_SECONDS / $2" | bc
+         )
+         
+         TIME_REMAINING=$(
+            echo "($2 - $PROGRESS_SECONDS) / $PROGRESS_SPEED" | bc
+         )
+         
+         update_status_field $1 'encode_remaining_sec' $TIME_REMAINING
+         update_status_field $1 'encode_progress' $PERCENT_COMPLETE
       fi
-   done
+   done < $TRANSCODE_PROGRESS_PIPE
 }
 
-echo "Initialising ffmpeg monitor script"
-
-mkfifo transcode-progress
-ssh root@$SERVER_ADDRESS '~/ffmpeg-monitor-progress.sh' &
+mkfifo $TRANSCODE_PROGRESS_PIPE
+ssh root@$SERVER_ADDRESS '~/ffmpeg-monitor-progress.sh' &>$TRANSCODE_PROGRESS_PIPE &
 
 # Copy each file, transcode and copy back
 for file in "${@:2}"
@@ -131,49 +139,50 @@ do
       $STATUS_FILE
    )
    
-   update_progress $DOWNLOADED_URL &
+   VIDEO_DURATION=$(
+      ffprobe -v error -show_entries format=duration -of default=noprint_wrappers=1:nokey=1 "$1/$file"
+   )
    
-   echo "Copying $1/$file to raw-$file..."
+   update_progress $DOWNLOADED_URL "$VIDEO_DURATION" &
+   
+   echo "Copying file $1/$file to root@$SERVER_ADDRESS:~/raw-$file"
    scp "$1/$file" root@$SERVER_ADDRESS:~/"raw-$file"
-   
-   echo "Running ffmpeg on 'raw-$file', output to '$file-hls'"
    
    # Spawn a separate process on the server to read the progress
    # of ffmpeg, which is piped to ~/ffmpeg-progress and then read by
    # ~/progress.sh, which emits its contents back to us 
    
+   echo "Running ffmpeg..."
    ssh root@$SERVER_ADDRESS "~/create-hls-stream.sh 'raw-$file' '$file-hls'" &>/dev/null
    
    echo "Copying file back..."
    scp -r root@$SERVER_ADDRESS:~/"$file-hls" $1
    
-   echo "Updating status.json..."
    # Update the status file
+   PLAYBACK_URL="$1/$file-hls/playlist.m3u8"
    
-   echo "Found downloaded URL: '$DOWNLOADED_URL'"
-   echo "Setting to complete..."
    flock $STATUS_FILE.lock \
    jq --arg url "$DOWNLOADED_URL" \
       --arg name "$file" \
+      --arg playback_url "$PLAYBACK_URL" \
       '.completed |= . + [{
          download_name: $name,
-         url: $url
+         origin_url: $url,
+         playback_url: $playback_url
       }]' $STATUS_FILE | \
    sponge $STATUS_FILE
    
-   echo "Removing old values..."
    flock $STATUS_FILE.lock \
    jq --arg name "$file" \
       'del(.downloaded[] | select(.download_name == $name))' \
       $STATUS_FILE | \
    sponge $STATUS_FILE
    
-   echo "HLS transcode of $file complete"
+   echo "Transcode of $file complete"
 done
 
 # Delete original media
 for file in "${@:2}"
 do
-   echo "Removing original file at $1/$file"
    rm "$1/$file"
 done
